Tidy up the start-stops API
Expose the two internal pieces, I can see situations where we might
already have start and stop sequences from somewhere else.

Be consistent about the startStops naming, I think this will be easier
to follow.
