+++ lib/index.js
@@ -3,6 +3,6 @@
  * Copyright 2023 AMRC
  */
 
-export * from "./items-to-observable.js";
 export * from "./k8s-watcher.js";
+export * from "./starts-stops.js";
 export * from "./util.js";
